Remove versions and MS/System packages from NOTICE (#40620)
We do not need to indicate that we consume System or Microsoft packages;
it is expected that we do so because we are Microsoft and we are using
.NET.

We also don't need to maintain a second list of package versions that is
bound to fall out of date.

We absolutely do not need to cause build breaks when those package
versions change because the build machine updated.

Closes #23321 (by alternative construction)