Window is clickthrough, wont show up as an app on Task manager
Before, the user could not click through the overlay window. Moreover, the overlay could be seen as an application on task manager, allowing the user to maximize/minimize it which does not make sense for an overlay. These changes make it so it only shows up as a process and the .exe for screencast mode does not show up on task manager